requirements.txt
dotenv-cli==3.0.1
speedtest-cli==2.1.3
psycopg2==2.9.3
